Split oversized team_admin module
